Update GEOtop snow parameters based on SnowDepthCalibration branch
The parameters are the results of a 1000-iteration randomized
calibration using a Latin Hypercube Sampling design which varied
SnowCorrFactor (multiplication factor for snow), ThresTempRain
(temperature above which all precipitation is rain), and ThresTempSnow
(temperature below which all precipitation is snow). This calibration
was run using NARR input data and without time-varying vegetation.