Update heat bottom boundary condition
The bottom boundary condition for heat transport is defined using a zero
amplitude temperature depth (20 m) and a temperature at this depth (-4.8
C). These values were obtained from the nearby Seabee borehole, which is
part of the Global Terrestrial Network for Permafrost, at:
http://gtnpdatabase.org/boreholes/view/812

This measurement is from 2007. Note that it is not explicitly stated
that 20 m is the zero amplitude temperature depth, but the standard for
taking ground measurements is using the zero amplitude temperature
depth, so I took this as a rough estimate. It doesn't have a big impact
on the simulation results, anyways.