Add script to resample daily meteo data to hourly
This uses the same relationships as AgroIBIS. Tmean is downscaled based
on Tmax/Tmin and solar time, from Campbell & Norman. Radiation is based
on the ratio of hourly to daily top of atmosphere solar radiation. Wind
is random. Rain generates an event length and a random start time.
Relative humidity is constant throughout the day.

Note that GEOtop is still set up to use daily files.